# dist/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "RESTful API",
        "description": "Backend RESTful API  with all CRUD operations using NodeJS, MongoDB and Redis.",
        "contact": {
            "email": "alex.buaiscia@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:3000/"
        },
        {
            "url": "https://restful-db-api.herokuapp.com/"
        }
    ],
...
    "paths": {
        "/movies/": {
            "get": {
                "tags": [
                    "movies"
                ],
                "summary": "Get all movies",
                "description": "Returns all movies",
                "operationId": "getAllMovies",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Movie"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error",
                        "content": {}
                    }
                }
            },
            "post": {
                "tags": [
                    "movies"
                ],
                "summary": "Add a new movie to the database",
                "operationId": "addMovie",
                "requestBody": {
                    "description": "Movie object that needs to be added to the database",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Movie"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "405": {
                        "description": "Invalid input",
                        "content": {}
                    }
                },
                "x-codegen-request-body-name": "body"
            }
        },
...
    },
    "components": {
        "schemas": {
            "Movie": {
                "type": "object",
                "properties": {
                    "movieId": {
                        "type": "string"
                    },
                    "title": {
                        "minLength": 1,
                        "type": "string"
                    },
                    "director": {
                        "minLength": 2,
                        "type": "string"
                    },
                    "description": {
                        "maxLength": 600,
                        "minLength": 20,
                        "type": "string"
                    },
                    "shortDescription": {
                        "maxLength": 150,
                        "minLength": 10,
                        "type": "string"
                    },
                    "duration": {
                        "minimum": 10,
                        "type": "number"
                    },
                    "releaseDate": {
                        "type": "string",
                        "format": "date"
                    },
                    "images": {
                        "type": "object",
                        "properties": {
                            "cover": {
                                "type": "string"
                            },
                            "poster": {
                                "type": "string"
                            },
                            "background": {
                                "type": "string"
                            }
                        }
                    },
                    "genre": {
                        "type": "string",
                        "enum": [
                            "Documentary",
                            "Scifi",
                            "Thriller",
                            "Biography",
                            "Romance",
                            "Children",
                            "Other"
                        ]
                    },
                    "childrenFriendly": {
                        "type": "boolean"
                    }
                }
            },
...
        }
    }
}